@@ -1,15 +1,16 @@
+'use client';
 import React, { useState, useEffect } from 'react';
 import { PlayerDetails } from '@/interfaces/players';
-import getStandings from '@/utils/getRankings';
+import apiHelper from '@/utils/apiHelper';
 import { skeletonTable } from '@/utils/skeletonTableHome';
 
 export default function StandingsTable() {
-  const [standings, setStandings] = useState([]);
+  const [standings, setStandings] = useState<PlayerDetails[]>([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     async function fetchStandings() {
-      const standingsData = await getStandings();
+      const standingsData = await apiHelper('standings');
       setStandings(standingsData);
       setLoading(false); // Data received, set loading to false
     }
@@ -18,12 +19,15 @@ export default function StandingsTable() {
   }, []);
 
   return (
-    <div className='flex flex-col justify-center items-center pt-2 '>
-      <div className='w-[337px] sm:w-[450px] bg-gradient-to-r from-cyan-600 to-blue-500 p-8 rounded-lg shadow-2xl'>
+    <div className='flex flex-col justify-center items-center'>
+      <h1 className='text-[#310639] text-2xl pb-5 font-semibold animate-fade-up'>
+        League Standings
+      </h1>
+      <div className='w-[337px] sm:w-[450px] bg-gradient-to-r from-cyan-600 to-blue-500 p-8 rounded-lg shadow-2xl border-2 border-black'>
         {loading ? (
           skeletonTable
         ) : (
-          <table className={'text-white w-[280px] sm:w-[400px] font-light text-sm'}>
+          <table className={'text-white w-[280px] sm:w-[400px] font-light text-sm '}>
             <thead>
               <tr className='border-b-2 border-white'>
                 <th className='font-medium w-1/4 py-2'>Player</th>
